Switched to Schema versioning
Before, we had a notion of the Latest and the Prior schemas, connoting
the schema we expect now versus the schema we expect prior to ever
running a migration. Now, we just have a version 0, and the expectation
is that someone writing a further application will implement the version
1, 2, 3, etc., and include their migrations between them in each of
those files.